+++ src/test/java/app/bottlenote/review/controller/MyReviewReadControllerTest.java
@@ -20,7 +20,7 @@ import app.bottlenote.review.domain.constant.ReviewDisplayStatus;
 import app.bottlenote.review.domain.constant.ReviewSortType;
 import app.bottlenote.review.domain.constant.SizeType;
 import app.bottlenote.review.dto.request.PageableRequest;
-import app.bottlenote.review.dto.response.ReviewDetail;
+import app.bottlenote.review.dto.response.ReviewListResponse;
 import app.bottlenote.review.dto.response.ReviewResponse;
 import app.bottlenote.review.service.ReviewService;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -126,7 +126,7 @@ class MyReviewReadControllerTest {
 	void test_case_1(String description, PageableRequest pageableRequest) throws Exception {
 
 		//given
-		PageResponse<ReviewResponse> response = getResponse();
+		PageResponse<ReviewListResponse> response = getResponse();
 
 		//when
 		when(SecurityContextUtil.getUserIdByContext()).thenReturn(Optional.of(userId));
@@ -155,7 +155,7 @@ class MyReviewReadControllerTest {
 	void test_fail_when_no_auth_info() throws Exception {
 
 		//given
-		PageResponse<ReviewResponse> response = getResponse();
+		PageResponse<ReviewListResponse> response = getResponse();
 
 		//when
 		when(SecurityContextUtil.getUserIdByContext()).thenReturn(Optional.empty());
@@ -227,9 +227,9 @@ class MyReviewReadControllerTest {
 			.andDo(print());
 	}
 
-	private PageResponse<ReviewResponse> getResponse() {
+	private PageResponse<ReviewListResponse> getResponse() {
 
-		ReviewDetail reviewDetail_1 = ReviewDetail.builder()
+		ReviewResponse reviewResponse_1 = ReviewResponse.builder()
 			.reviewId(1L)
 			.reviewContent("맛있습니다")
 			.price(BigDecimal.valueOf(100000L))
@@ -248,7 +248,7 @@ class MyReviewReadControllerTest {
 			.hasReplyByMe(false)
 			.build();
 
-		ReviewDetail reviewDetail_2 = ReviewDetail.builder()
+		ReviewResponse reviewResponse_2 = ReviewResponse.builder()
 			.reviewId(2L)
 			.reviewContent("나름 먹을만 하네요")
 			.price(BigDecimal.valueOf(110000L))
@@ -268,7 +268,7 @@ class MyReviewReadControllerTest {
 			.build();
 
 		Long totalCount = 2L;
-		List<ReviewDetail> reviewDetails = List.of(reviewDetail_1, reviewDetail_2);
+		List<ReviewResponse> reviewResponses = List.of(reviewResponse_1, reviewResponse_2);
 		CursorPageable cursorPageable = CursorPageable.builder()
 			.currentCursor(0L)
 			.cursor(1L)
@@ -276,7 +276,7 @@ class MyReviewReadControllerTest {
 			.hasNext(false)
 			.build();
 
-		ReviewResponse response = ReviewResponse.of(totalCount, reviewDetails);
+		ReviewListResponse response = ReviewListResponse.of(totalCount, reviewResponses);
 		return PageResponse.of(response, cursorPageable);
 	}
 
@@ -286,7 +286,7 @@ class MyReviewReadControllerTest {
 	void test_sortType(String sortType, int expectedStatus) throws Exception {
 
 		// given
-		PageResponse<ReviewResponse> response = getResponse();
+		PageResponse<ReviewListResponse> response = getResponse();
 
 		// when
 		when(SecurityContextUtil.getUserIdByContext()).thenReturn(Optional.of(userId));
@@ -313,7 +313,7 @@ class MyReviewReadControllerTest {
 	void test_sortOrder(String sortOrder, int expectedStatus) throws Exception {
 
 		// given
-		PageResponse<ReviewResponse> response = getResponse();
+		PageResponse<ReviewListResponse> response = getResponse();
 
 		// when
 		when(SecurityContextUtil.getUserIdByContext()).thenReturn(Optional.of(userId));
